+++ src/components/QuizQuestions/QuizQuestions.scss
@@ -1,0 +1,8 @@
+.quizquestions{
+    &_section{
+        padding: 3rem;
+        display:flex;
+        flex-direction: column;
+        gap:2rem;
+    }
+}
