+++ src/pages/QuizPage/QuizPage.scss
@@ -37,9 +37,6 @@
       display:flex;
       flex-direction: column;
       gap:2rem;
-      &--field {
-       // width: 90% !important;
-      }
     }
   }
 }
